refactor: centralize ElasticSearch to elasticapi
